requirements.txt
click>=8.1.0
rich>=13.0.0
pathspec>=0.12.0
openai>=1.0.0
pytest>=8.0.0
python-dotenv>=1.0.0
